refactor: split enums and classes into separate files
