+++ Sources/DesktopManager/DesktopManagerWrapper.cs
@@ -1,0 +1,8 @@
+namespace DesktopManager;
+
+/// <summary>
+/// Wrapper class for the desktop manager.
+/// </summary>
+[ComImport, Guid("C2CF3110-460E-4fc1-B9D0-8A1C0C9CC4BD")]
+internal class DesktopManagerWrapper {
+}
