+++ Sources/DesktopManager/DesktopSlideShowOptions.cs
@@ -1,0 +1,11 @@
+namespace DesktopManager;
+
+/// <summary>
+/// This enumeration is used to set and get slide show options.
+/// </summary>
+public enum DesktopSlideShowOptions {
+    /// <summary>
+    /// When set, indicates that the order in which images in the slide show are displayed can be randomized.
+    /// </summary>
+    ShuffleImages = 0x01
+}
