+++ Sources/DesktopManager/DesktopSlideshowDirection.cs
@@ -1,0 +1,15 @@
+namespace DesktopManager;
+
+/// <summary>
+/// This enumeration is used by the AdvanceSlideshow method to indicate whether to advance the slide show forward or backward.
+/// </summary>
+public enum DesktopSlideshowDirection {
+    /// <summary>
+    /// Advance the slideshow forward.
+    /// </summary>
+    Forward = 0,
+    /// <summary>
+    /// Advance the slideshow backward.
+    /// </summary>
+    Backward = 1,
+}
