+++ Sources/DesktopManager/DisplayOrientation.cs
@@ -1,0 +1,23 @@
+namespace DesktopManager;
+
+/// <summary>
+/// Specifies the orientation of the display.
+/// </summary>
+public enum DisplayOrientation {
+    /// <summary>
+    /// Default landscape orientation.
+    /// </summary>
+    Default = 0,
+    /// <summary>
+    /// Rotated 90 degrees.
+    /// </summary>
+    Degrees90 = 1,
+    /// <summary>
+    /// Rotated 180 degrees.
+    /// </summary>
+    Degrees180 = 2,
+    /// <summary>
+    /// Rotated 270 degrees.
+    /// </summary>
+    Degrees270 = 3
+}
