+++ Sources/DesktopManager/IDesktopManager.cs
@@ -115,10 +115,3 @@ public interface IDesktopManager {
     /// <returns>True if enabled, otherwise false.</returns>
     bool Enable();
 }
-
-/// <summary>
-/// Wrapper class for the desktop manager.
-/// </summary>
-[ComImport, Guid("C2CF3110-460E-4fc1-B9D0-8A1C0C9CC4BD")]
-internal class DesktopManagerWrapper {
-}
