+++ Sources/DesktopManager/SystemTheme.cs
@@ -1,0 +1,15 @@
+namespace DesktopManager;
+
+/// <summary>
+/// Represents the Windows color theme.
+/// </summary>
+public enum SystemTheme {
+    /// <summary>
+    /// Light mode.
+    /// </summary>
+    Light = 0,
+    /// <summary>
+    /// Dark mode.
+    /// </summary>
+    Dark = 1
+}
