+++ Sources/DesktopManager/WindowInfo.cs
@@ -76,9 +76,3 @@ public class WindowInfo {
     /// </summary>
     public WindowState? State { get; set; }
 }
-
-/// <summary>
-/// Represents the position and size information of a window.
-/// </summary>
-public class WindowPosition : WindowInfo {
-}
