+++ Sources/DesktopManager/WindowMessage.cs
@@ -1,0 +1,11 @@
+namespace DesktopManager;
+
+/// <summary>
+/// Window message constants.
+/// </summary>
+public enum WindowMessage {
+    /// <summary>
+    /// System command message.
+    /// </summary>
+    WM_SYSCOMMAND = 0x0112
+}
