+++ Sources/DesktopManager/WindowPosition.cs
@@ -1,0 +1,7 @@
+namespace DesktopManager;
+
+/// <summary>
+/// Represents the position and size information of a window.
+/// </summary>
+public class WindowPosition : WindowInfo {
+}
